requirements.txt
playwright==1.49.0
playwright-stealth==1.0.6
pyotp==2.9.0
openai==1.55.0
PyYAML==6.0.1
python-dotenv==1.0.1
httpx<0.28.0
